+++ includes/rts/Flags.h
@@ -191,6 +191,7 @@ typedef struct _MISC_FLAGS {
     bool install_signal_handlers;
     bool install_seh_handlers;
     bool generate_dump_file;
+    bool generate_stack_trace;
     bool machineReadable;
     StgWord linkerMemBase;       /* address to ask the OS for memory
                                   * for the linker, NULL ==> off */