+++ rts/linker/PEi386.c
@@ -153,6 +153,10 @@
 static uint8_t* cstring_from_COFF_symbol_name(
     uint8_t* name,
     uint8_t* strtab);
+#include <inttypes.h>
+#include <dbghelp.h>
+#include <stdlib.h>
+#include <Psapi.h>
 
 #if defined(x86_64_HOST_ARCH)
 static size_t makeSymbolExtra_PEi386(
@@ -2059,4 +2063,196 @@ SymbolAddr *lookupSymbol_PEi386(SymbolName *lbl)
     }
 }
 
+/* -----------------------------------------------------------------------------
+ * Debugging operations.
+ */
+
+pathchar*
+resolveSymbolAddr_PEi386 (pathchar* buffer, int size,
+                          SymbolAddr* symbol, uintptr_t* top ){
+    SYMBOL_INFO sym;
+    ZeroMemory (&sym, sizeof(SYMBOL_INFO));
+    sym.MaxNameLen = sizeof(char) * 1024;
+
+    DWORD64 uDisplacement = 0;
+    HANDLE hProcess = GetCurrentProcess();
+    ObjectCode* obj = NULL;
+    uintptr_t start, end;
+    *top = 0;
+
+    pathprintf (buffer, size, WSTR("0x%" PRIxPTR), symbol);
+
+    if (SymFromAddr (hProcess, (uintptr_t)symbol, &uDisplacement, &sym))
+    {
+      /* Try using Windows symbols.  */
+      wcscat (buffer, WSTR(" "));
+      pathchar* name = mkPath (sym.Name);
+      wcscat (buffer, name);
+      stgFree (name);
+      if (uDisplacement != 0)
+      {
+        int64_t displacement = (int64_t)uDisplacement;
+        pathchar s_disp[50];
+        if (displacement < 0)
+          pathprintf ((pathchar*)s_disp, 50, WSTR("-%ld"), -displacement);
+        else
+          pathprintf ((pathchar*)s_disp, 50, WSTR("+%ld"), displacement);
+
+        wcscat (buffer, s_disp);
+      }
+    }
+    else
+    {
+      /* Try to calculate from information inside the rts.  */
+      uintptr_t loc = (uintptr_t)symbol;
+      for (ObjectCode* oc = objects; oc; oc = oc->next) {
+          for (int i = 0; i < oc->n_sections; i++) {
+              Section section = oc->sections[i];
+              start = (uintptr_t)section.start;
+              end   = start + section.size;
+              if (loc > start && loc <= end)
+              {
+                  wcscat (buffer, WSTR(" "));
+                  if (oc->archiveMemberName)
+                  {
+                      pathchar* name = mkPath (oc->archiveMemberName);
+                      wcscat (buffer, name);
+                      stgFree (name);
+                  }
+                  else
+                  {
+                      wcscat (buffer, oc->fileName);
+                  }
+                  pathchar s_disp[50];
+                  pathprintf (s_disp, 50, WSTR("+0x%" PRIxPTR), loc - start);
+                  wcscat (buffer, s_disp);
+                  obj = oc;
+                  goto exit_loop;
+              }
+          }
+      }
+
+      /* If we managed to make it here, we must not have any symbols nor be
+         dealing with code we've linked. The only thing left is an internal
+         segfault or one in a dynamic library. So let's enumerate the module
+         address space.  */
+      HMODULE *hMods = NULL;
+      DWORD cbNeeded;
+      EnumProcessModules (hProcess, hMods, 0, &cbNeeded);
+      hMods = stgMallocBytes (cbNeeded, "resolveSymbolAddr_PEi386");
+      if (EnumProcessModules (hProcess, hMods, cbNeeded, &cbNeeded))
+      {
+        uintptr_t loc = (uintptr_t)symbol;
+        MODULEINFO info;
+        for (uint32_t i = 0; i < cbNeeded / sizeof(HMODULE); i++) {
+           ZeroMemory (&info, sizeof (MODULEINFO));
+           if (GetModuleInformation (hProcess, hMods[i], &info,
+                                     sizeof(MODULEINFO)))
+           {
+             uintptr_t start = (uintptr_t)info.lpBaseOfDll;
+             uintptr_t end   = start + info.SizeOfImage;
+             if (loc >= start && loc < end)
+             {
+                /* Hoera, finally found some information.  */
+                pathchar tmp[MAX_PATH];
+                if (GetModuleFileNameExW (hProcess, hMods[i], tmp, MAX_PATH))
+                {
+                  wcscat (buffer, WSTR(" "));
+                  wcscat (buffer, tmp);
+                  pathprintf (tmp, MAX_PATH, WSTR("+0x%" PRIxPTR), loc - start);
+                  wcscat (buffer, tmp);
+                }
+                break;
+             }
+           }
+        }
+      }
+
+      stgFree(hMods);
+    }
+
+    /* Finally any file/line number.  */
+    IMAGEHLP_LINE64 lineInfo = {0};
+    DWORD dwDisplacement = 0;
+  exit_loop:
+    if (SymGetLineFromAddr64(hProcess, (uintptr_t)symbol, &dwDisplacement,
+                             &lineInfo))
+    {
+      /* Try using Windows symbols.  */
+      pathchar s_line[512];
+      pathprintf ((pathchar*) s_line, 512, WSTR("   %ls (%lu)"),
+                  lineInfo.FileName, lineInfo.LineNumber);
+      wcscat (buffer, s_line);
+      if (dwDisplacement != 0)
+      {
+        pathprintf ((pathchar*) s_line, 512, WSTR(" +%lu byte%s"),
+                    dwDisplacement,
+                    (dwDisplacement == 1 ? WSTR("") : WSTR("s")));
+      }
+      wcscat (buffer, s_line);
+    }
+    else if (obj)
+    {
+      /* Try to calculate from information inside the rts.  */
+      typedef struct _SymX { SymbolName* name; uintptr_t loc; } SymX;
+      SymX* locs = stgCallocBytes (sizeof(SymX), obj->n_symbols,
+                                   "resolveSymbolAddr");
+      int blanks = 0;
+      for (int i = 0; i < obj->n_symbols; i++) {
+          SymbolName* sym = obj->symbols[i];
+          if (sym == NULL)
+            {
+               blanks++;
+               continue;
+            }
+          RtsSymbolInfo* a = NULL;
+          ghciLookupSymbolInfo(symhash, sym, &a);
+          if (a) {
+              SymX sx = {0};
+              sx.name = sym;
+              sx.loc  = (uintptr_t)a->value;
+              locs[i] = sx;
+          }
+      }
+      int comp (const void * elem1, const void * elem2)
+      {
+          SymX f = *((SymX*)elem1);
+          SymX s = *((SymX*)elem2);
+          if (f.loc > s.loc) return  1;
+          if (f.loc < s.loc) return -1;
+          return 0;
+      }
+      qsort (locs, obj->n_symbols, sizeof (SymX), comp);
+      uintptr_t key  = (uintptr_t)symbol;
+      SymX* res = NULL;
+
+      for (int x = blanks; x < obj->n_symbols; x++) {
+          if (x < (obj->n_symbols -1)) {
+              if (locs[x].loc >= key && key < locs[x+1].loc) {
+                res = &locs[x];
+                break;
+              }
+          }
+          else
+          {
+              if (locs[x].loc >= key) {
+                  res = &locs[x];
+                  break;
+                }
+          }
+      }
+
+      if (res) {
+          pathchar s_disp[512];
+          *top = (uintptr_t)res->loc;
+          pathprintf ((pathchar*)s_disp, 512,
+                      WSTR("\n\t\t (%s+0x%" PRIxPTR ")"),
+                      res->name, res->loc - key);
+          wcscat (buffer, s_disp);
+      }
+      stgFree (locs);
+    }
+
+    return buffer;
+}
 #endif /* mingw32_HOST_OS */